requirements.txt
appdirs==1.4.3
attrs==19.3.0
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.14.0
Click==7.0
distlib==0.3.0
filelock==3.0.12
Flask==1.1.1
Flask-Cors==3.0.8
flask-jsonschema-validator==0.0.4
Flask-JWT==0.3.2
Flask-JWT-Extended==3.24.1
flask-marshmallow==0.11.0
Flask-SQLAlchemy==2.4.1
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
MarkupSafe==1.1.1
marshmallow==3.5.0
marshmallow-sqlalchemy==0.22.3
pipenv==2018.11.26
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.9.3
pyrsistent==0.15.7
python-dotenv==0.12.0
six==1.14.0
SQLAlchemy==1.3.13
virtualenv==20.0.7
virtualenv-clone==0.5.3
Werkzeug==1.0.0
zipp==3.0.0
